+++ src/Commands/GenerateRulesCommand.php
@@ -14,7 +14,7 @@ use LaracraftTech\LaravelSchemaRules\Exceptions\ColumnDoesNotExistException;
 use LaracraftTech\LaravelSchemaRules\Exceptions\FailedToCreateRequestClassException;
 use LaracraftTech\LaravelSchemaRules\Exceptions\MultipleTablesSuppliedException;
 use LaracraftTech\LaravelSchemaRules\Exceptions\TableDoesNotExistException;
-use LaracraftTech\LaravelSchemaRules\Resolvers\SchemaRulesResolverInterface;
+use LaracraftTech\LaravelSchemaRules\Contracts\SchemaRulesResolverInterface;
 
 class GenerateRulesCommand extends Command
 {
